[mod] Added Copy feature for Sierra or Higher
MacOS auto-extracts ZIPs unless you use Chrome/Firefox or tell Safari not to open "safe" files in Preferences and uncheck the box. Due to this, MacOS Release of ML will require 2 additional inputs surrounding if the user downloaded DDLC off Chrome/Firefox or Safari with either Safe files On or Off and if the OS is El Capitan or lower or Sierra or Higher (Sierra -> Mojave ATM as Catalina Beta has issues) before extracting or copying.

When shutil.copytree is in effect, it will only copy the files in the folder and not the folder itself (e.g. Mod Folder name is TTR but temp folder gives out only the game folder or .rpa/rpyc's. This has been resolved now, but to keep note to relabel a bit.