[mod] Removed OSX Feature. Final Build of 5.1.0
As stated by @alicerunsonfedora, Mac OS and OS X had extract features that date back to 2008 around OS X 10.5 Snow Leopard on Safari. Due to this the OS question has been removed and only asks the user if they are using Safari with Safe Files On or Off OR Chrome/Firefox/Opera.